+++ src/features/export-package-resources-modal/export-package-resources-modal.js
@@ -10,6 +10,8 @@ import {
   Modal,
   ModalFooter,
 } from '@folio/stripes/components';
+import { useCallout } from '@folio/stripes/core';
+
 
 import ExportFieldsSection from './export-fields-section';
 import { useExportPackageTitle } from '../../hooks';
@@ -35,6 +37,7 @@ const ExportPackageResourcesModal = ({
   titleSearchFilters,
 }) => {
   const intl = useIntl();
+  const callout = useCallout();
   const [fieldSectionState, setFieldSectionState] = useState({
     [RECORD_TYPES.PACKAGE]: {
       allSelected: true,
@@ -47,8 +50,22 @@ const ExportPackageResourcesModal = ({
   });
 
   const { doExport } = useExportPackageTitle({
-    onError: () => {},
-    onSuccess: () => {},
+    onSuccess: (res) => {
+      callout.sendCallout({
+        type: 'success',
+        message: intl.formatMessage({
+          id: 'ui-eholdings.exportPackageResources.toast.success',
+        }, {
+          jobName: res.name,
+        }),
+      });
+    },
+    onError: () => {
+      callout.sendCallout({
+        type: 'error',
+        message: intl.formatMessage({ id: 'ui-eholdings.exportPackageResources.toast.fail' }),
+      });
+    },
   });
 
   const canExport = useMemo(() => {